requirements.txt
python-telegram-bot
requests
asyncio
python-dotenv
tweepy

instagrapi
pillow
